[utilities] let sos_get_command_output() write directly to files
Add a 'dest' kwarg to sos_get_command_output() that if set is a
writable file-like object, to which output will be directed,
instead of the in-memory deque.

This can be used by callers of the function to avoid possible
memory errors when handling commands that produce very large
output (e.g. journalctl).

Currently no size limiting is implemented for direct written
output. This is difficult to implement in-line due to the
uncertainty of exactly how much data will be written by the
command, but could be implemented as a post-collection operation
by unlinking and re-writing the file once the size is known.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>