[juju] Support for Juju version 2
Adding support for Juju 2.x.  This plugin needed some cleanup.  There are 3
use cases for the juju plugin: the juju client, a juju deployed machine, and
a juju controller.  I have broken the plugin into 3 parts to reflect this.

The juju plugin is now specifically for the juju client use case.  Juju is
typically installed as a snap or a package, therefore I added /snap/bin to the
path for Debian to handle this until a better method for handling snaps is
figured out.  Juju is configured as a user to communicate with the controller
it is therefore necessary to run the juju commands with sudo for them to
execute properly.  This plugin now collects as much information about the juju
environment as possible by using the juju command, this information includes
the configured clouds, models, spaces, model configs, status of each model,
and configuration of each app in each model.

The juju_machine plugin is for the machines deployed by juju use case.  This
plugin handles config, log and journal collection for a juju machine.  There
is no package installed to indicate it is a juju machine, so this plugin is
activated by the presence of /var/log/juju.

The juju_controller plugin is for the juju controller use case.  A juju
controller is a juju machine so the machine plugin should handle the
collection of logs from that aspect.  The only identifying piece of
information that this is a juju controller is the /var/lib/juju/db directory
and really juju is just a mongodb server.  This plugin instruments some mongo
queries to get the status of that database.

Signed-off-by: Nick Niehoff nick.niehoff@canonical.com