[insights] fix handling of configs tuple
The configs member is now a tuple: we can't just slam it into a
add_copy_specs() list or a do_file_sub() call and have it work.

Split out the a_c_s() and pass the tuple alone (it looks less ugly
than having a cons inside the list), and loop over the elements in
postproc to handle the possible file locations.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>