[systemd] avoid collecting /dev/null
Avoid "collecting" /dev/null as part of the systemd plugin.

In my testing that was the only plugin that brought it in.

Creating these character devices makes it harder to manage
(delete) extracted sosreports without more permissions.

This supersedes #1469 which removed the ability to "collect"
char devices.  As we only have one example today, let's just
exclude that.

Signed-off-by: Bryan Quigley <bryan.quigley@canonical.com>