[sosreport] handle non-valued and counter options in sos.conf
Command line arguments that are non-valued ("store_true"), or that
have counter semantics need special handling when read from the
configuration file (since optionless values *do* have an explicit
value in the file, in order to meet .ini file structure, and since
counter values are represented in the file as an integer).

Process these options specially and return them in a form that is
correctly parsed by the ArgumentParser.

Related: #1530

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>