[unpackaged] check plugin predicate before processing packages
The unpackaged plugin is unusual in that it does a lot of costly
processing in Python, without ever calling out to an external
binary. This means that the normal, automatic predicate testing
hooks are not executed until we come to add the final string to
the archive (by which time we've already done all the work).

Make a special case for this plugin, and test the current plugin
predicate before operating on package lists.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>