[Plugin] Terminate running commands when a plugin exceeds timeout
Previously, if a plugin timeout was hit the thread it was running in
would be destroyed, however if it were running a command (ultimately
handled by a Popen() call), that command would still continue to run
under PID 1. This could then cause sos to report I/O exceptions after
the archive has been closed with an attempt to write to the temporary
directory used for the archive.

Now, utilities.sos_get_command_output() accepts (and is handed from
Plugin) a polling method that is used to monitor if the timeout has been
hit and will call terminate() to end kill the running command if that
polling method returns True.

For commands run outside of a plugin, previous behavior remains the
same since now AsyncReader.get_contents() will block until the command
completes (regardless of exit code) rather than relying on the join()
when an AsyncReader was initialized as was done before.

Resolves: #1590

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>