[Plugin] handle predicates correctly in get_cmd_output_now()
The get_command_output_now() method is used for two distinct
purposes:

 * Called directly from user plugin code, to immediately collect
   the output of a command during the setup() phase (e.g. to then
   process the output to inform further collection, while also
   storing the result in the report).

 * Internally from the _collect_cmd_output() callback, in order
   to collect the commands added at setup() time by user plugins
   during the collect() phase.

In the former case, predicates should be applied as normal and
evaluated prior to command execution. In the latter, predicates
have already been applied at the time of the add_cmd_output()
call; any predicate that was not set at the time of that call
must not apply to the collection of the command.

Split the method in two: _g_c_o_n() now collects command output
unconditionally, and is used directly by internal callers, and
g_c_o_n() becomes a wrapper to perform predicate evaluation.

Resolves: #1616

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>