[kubernetes] Allow only capturing logs for certain pods
The 'podlogs' option was previously an all-or-nothing toggle, which
could result in incredibly large amounts of data collection when only a
few pod's logs were needed.

This change adds a 'podlogs-filter' option that allows users to provide
either a specific pod name or a regex to match pod names against, to
limit the amount of pod logs we collect.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>