[gluster] Add volume, state information and dump glusterd processes
Adds collection of 'gluster volume get $vol all' output for all volume
names as well as the generate dumpfile from 'gluster get-state' when the
'dump' option is used. Note that this file will be under
/var/run/gluster in the archive.

Adds 'glusterd' to the list of processes sent signal USR1 when the
'dump' option is used. Removed the prep work previously done to redirect
where the statedump files are written to, as it has been reported that
those attempts do not function as intended. Statedumps are now collected
from /var/run/gluster.

Finally, some re-organization to ease future maintenance and updating
the size-limiting logic to use the implicit limiting in add_copy_spec().

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>