[containers_log] Fetch all container logs
Until now, we didn't have a central plugin for collecting
service container logs. Each service was managing its own
things, and this lead to missing log files in the past.

Ensuring we fetch the content of /var/log/containers on its
own ensures we won't forget to add possible new directories,
like the "stdouts" we recently added in OpenStack in order
to get the container STDOUT written on the disk.

Therefore, this commit does 2 main things:
- cleanup all the /var/logs/containers custom copy
- create a central plugin that takes care of this location on
  its own. It's active by default and ensure the location
  actually exists.

Resolves: #1668

Signed-off-by: Cédric Jeanneret <cjeanner@redhat.com>
Signed-off-by: Pavel Moravec <pmoravec@redhat.com>