[juju] Add support for Juju 2.x / refactoring work
Adding support for Juju 2.x. This plugin needed some cleanup.
There are 3 use cases for the juju plugin: the juju client,
a juju deployed machine, and a juju controller.
We have broken the plugin into 3 parts to reflect this.

The juju plugin is now specifically for the juju client use case.
Juju is typically installed as a snap or a package, therefore, we added
/snap/bin to the path for Debian to handle this until a better method
for handling snaps is figured out.

Juju is configured as a user to communicate with the controller it is
therefore necessary to run the juju commands with sudo for them to
execute properly.

This plugin now collects as much information about the juju environment
as possible by using the juju command, this information includes the
configured clouds, models, spaces, model configs, status of each model
and configuration of each app in each model.

Closes #1653

Co-authored-by: David Negreira <david.negreira@canonical.com>
Co-authored-by: Nick Niehoff <nick.niehoff@canonical.com>
Signed-off-by: David Negreira <david.negreira@canonical.com>