[Plugin] Add method to scrub certs/keys from collected files
This adds a new do_file_private_sub() method to the Plugin class to
allow for easy, and standardized scrubbing of private data such as
certificates or key pairs from files collected by sos, in much the same
way that do_cmd_private_sub() provides that functionality for command
output.

This method only takes a path or path regex representing the
filename/pattern to match against. Matches are replaced with
"-----SCRUBBED $thing" where $thing is the specific type of content
being replaced as provided in the collected file content. For example, a
private key would be replaced with "-----SCRUBBED RSA PRIVATE KEY-----".

This is done so that support representatives can still know the type of
information that was collected, without knowing the actual sensitive
content.

Resolves: #1690

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>