[openstack_*] Remove duplicate rpm verify from OpenStack plugins
A number of openstack plugins were calling 'rpm -V' against their own
package lists as part of their setup() method when the --verify option
is given to sos. This is already done in normal plugin execution, so
there is no need to duplicate the calls.

This change removes these calls from the openstack plugins.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>