[archive] Handle checking container sysroot in _make_leading_paths
Previously, in _make_leading_paths(), checking host file paths did not
account for non / sysroots, for situations where sos is run in a
container and the host's / is actually mounted under /host for example.
This would lead to copy errors when trying to copy symlinks.

This method now will use sysroot if one is set, thus avoiding copy
errors.

Credit to Pavel Moravec for identification of where this flaw was
originating, and the correction within _make_leading_paths().

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>