[Plugin] Add get_cmd_content_now method
Adds a new get_cmd_content_now() method that functions the same as
get_cmd_output_now() but will instead return the output of the command
instead of just the filename. The returned data is the same standard
get_command_output() dict with an additional 'filename' element.

The main purpose of this new method is to alleviate the trend in plugins
using get_cmd_output_now() and then manually opening the resulting file
in the archive for further parsing.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>