[process] Specify 20 for wchan length
wchan column is now truncated because there are many columns.
It is impossible to distinguish the symbol name if there are many
similar symbol names. Therefore, specify wchan size.

In this commit, we choose 20 for the wchan size but there could be
better value. Let's change this if it turns out that in the future.

Example: here is an output of ps command when plymouthd process
got hanged:

    F  UID    PID PPID PRI  NI    VSZ   RSS WCHAN  STAT TTY      TIME COMMAND
    5    0    159    1  20   0  44644  1696 n_tty_ S    ?        0:04 /bin/plymouthd --attach-to-session

The prefix n_tty_ is used in tty device driver of linux kernel
as follows:

    struct tty_ldisc_ops tty_ldisc_N_TTY = {
	    .magic           = TTY_LDISC_MAGIC,
	    .name            = "n_tty",
	    .open            = n_tty_open,
	    .close           = n_tty_close,
	    .flush_buffer    = n_tty_flush_buffer,
	    .chars_in_buffer = n_tty_chars_in_buffer,
	    .read            = n_tty_read,
	    .write           = n_tty_write,
	    .ioctl           = n_tty_ioctl,
	    .set_termios     = n_tty_set_termios,
	    .poll            = n_tty_poll,
	    .receive_buf     = n_tty_receive_buf,
	    .write_wakeup    = n_tty_write_wakeup,
	    .fasync          = n_tty_fasync,
    };

As you can see, it is impossible to find the symbol name of the
kernel function where plymouthd process slept from n_tty_.

Resolves: #1722

Signed-off-by: Tomofumi Yoshida <yoshida.tomo@fujitsu.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>