[openstack_instack] Add container-image-prepare log
During undercloud and overcloud deploy preparation steps, a task is
fetching container images and pushing them to a local registry
(usually on the undercloud directly).

Since this task might fail at some point, we need this log in order to
properly debug the issue and provide better support.

The task itself doesn't produce any other output than this log, and only
exists on the undercloud.

Resolves: #1730

Signed-off-by: Cédric Jeanneret <cjeanner@redhat.com>
Signed-off-by: Pavel Moravec <pmoravec@redhat.com>