[utilities] Fix high CPU usage and slow command collection
commit fc6721ac83c2 ("[Plugin] Terminate running commands when a plugin
exceeds timeout") introduced a polling method to sos_get_command_output()
but it is busy wait, and seems to increase the CPU usage and slow down
AsyncReader significantly.

As a result, a command that outputs much data like journalctl takes
longer time, and the plugin times out.

Inserting a sleep is a possible fix for this.

Signed-off-by: Kazuhito Hagio <k-hagio@ab.jp.nec.com>