[Plugin] add generic log_skipped_command() handler
Add a generic method to Plugin to log the reason that a command
was skipped during predicate evaluation. The method accepts bool
arguments indicating whether the command relies on kmod or service
presence checks, and a 'change' argument that indicates that it
may alter system configuration (adn so can be collected if the
user supplies --allow-system-changes).

Related: #1678

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>