[satellite] remove plugin for spacewalk
Spacewalk has and rhn has been replaced in favour of
foreman/katello/candlepin/pulp. Currently the only usage of the plugin
is on Satellite5 on RHEL6 and nowhere else. Other distros dont use it,
spacewalk upstream does not use sosreport, so the removal wont affect anybody.

And last but not least, the plugin name is confusing as it has nothing
in common with Satellite6.

Resolves: #1775

Signed-off-by: Pavel Moravec <pmoravec@redhat.com>