[kernel] Add the output of 'dmesg -H'
At the moment we have the output of 'dmesg' only in the sosreport, plus
the file /var/log/dmesg. But the output of 'dmesg' doesn't contain
timestamps in a human readable format. While the timestamps could be
inaccurate in some circumstances (see the man page of dmesg), it
can be still very helpful.

Signed-off-by: Jose Castillo <jose.mfcastillo@gmail.com>