[python] Add hashes option, refactor Python vs. RedHatPython
Adds a new `hashes` option to the `python` plugin. When specified, this
option will generate a hash for every python file discovered in the
`/usr/lib*` and `/usr/local/lib*` canonical paths. This information is
then written to the archive as `digests.json`

Additionally, refactors the differences between the base `Python` and
`RedHatPython` classes when dealing with the system-default python
installation.

Resolves: #1801

Signed-off-by: Bissenbay Dauletbayev <bissenbay@gmail.com>
Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>