[Plugin] Copy empty directories _do_copy_paths
Empty directories too provide some valid and helpful information in the
path copied, for example, Kernel livepatching subsystem exports the
functions being livepatched as empty directories instead of a file.
Similarly disable directory of apparmor, which indicates the status of
the profile.

The current code copies a directory to the archive only if it contains
files and skips in the case of an empty directory. This patch, copies
such skipped directory too.

Signed-off-by: Kamalesh Babulal <kamalesh@linux.ibm.com>