[redhat] Add missing plugins and options to 'cantboot' preset
The cantboot preset's initial plugins is missing several useful
plugins.
In addition there is several plugin options that should be included,
that are asked for separately on most can't boot cases.
The rpmdb option will increase the size of the sosreport, but would
be extremely helpful in a lot of can't boot cases.

Signed-off-by: Ryan Blakley <rblakley@redhat.com>