[podman] Collect network information and cert/hooks list
First, adds collection of a recursive listing of /etc/containers to show
which certs and oci hooks may be present.

Second, adds collection of podman networking information. Both the
output of 'podman network ls' and a subsequent inspect on each network
reported are captured. The inspect outputs are saved to the 'networks'
subdir.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>