[Plugin] Avoid potential duplicate sysroot join
An issue was discovered when sos was run in a container where the
sysroot when set to something other than '/' (e.g. '/host' in the Red
Hat support-tools container) was getting duplicated in checks against
defined forbidden paths.

This was because self.join_sysroot() is called in multiple spots during
file collection phases (both in `setup()` and `collect()`) before it is
additionally called in the internal `_is_forbidden_path()`. Since the
test for if `join_sysroot()` should be called will always return True in
a containerized environment (because the sysroot is not '/'), this would
cause the container's sysroot (e.g. '/host') to be appended again,
creating for example '/host/host/foo.txt'. Since this duplicated sysroot
append would cause forbidden path comparisons to fail, sos would not
only collect blacklisted sensitive files but would also hang in the
`kernel` plugin due to the fact it was now trying to collect trace_pipe
and other such files that it should otherwise skip.

This commit adds a conditional to `_is_forbidden_path()` to not call
`join_sysroot()` if the path in question already starts with the
sysroot.

Resolves: #1842

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>