[openstack_horizon] fix useless forbidden paths
The code has currently the following line:

        self.add_forbidden_path("*.py[co]")

But add_forbidden_path() needs a directory prefix, so this won't
work. This line was introduced via commit 4d75385ad to solve
issue #856, that discussed the postproc method in
openstack_horizon. It seems that the python object and source
files are present in /etc/openstack-dashboard/local_settings.d/
used to customize the dashboard. This patch adds that directory
to the add_forbidden_path() line.

Closes: #1846

Signed-off-by: Jose Castillo <jose.mfcastillo@gmail.com>