[Policy] Add Container Runtime abstraction class
Adds a `ContainerRuntime()` class to allow policies to specify a
container runtime to allow plugins to utilize.

The `ContainerRuntime` is intended to allow for the discovery of
containers, including specific ones by name, and for the execution of
commands inside those containers.

This is meant to remove the overhead of manually defining ways to
determine an active runtime and if a component is containerized within
plugins.

Related: #1866

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>