[Plugin] Integrate ContainerRuntime in Plugin API
Integrates the ContainerRuntime checks to the Plugin API -
`Plugin.exec_cmd()` now accepts a `container` argument that, if set,
will first check to see if the container exists and if it does, format
the requested command to be executed by the runtime in the selected
container. If the runtime is not available, or the container does not
exist, the command is not run.

Additionally adds a `container_exists()` call to check to see if a
container is present/running, and a `fmt_container_cmd()` call that
wraps the `ContainerRuntime` method to allow for easy formatting of
commands for `add_cmd_output()` calls.

Closes: #1866

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>