[plugins] Support for directory specification in add_copy_spec()
Previously, when a directory was specified in add_copy_spec(),
stat.ST_SIZE returned the size of the directory.
As a result, the limitation by --log-size did not work.

When a directory is specified in add_copy_spec(),
use os.walk() to find out the files under the directory.
From this, the limitation by --log-size work correctly.

Signed-off-by: MIZUTA Takeshi <mizuta.takeshi@fujitsu.com>