[pulp] Fix invalid syntax in collected json files
Plugin replaces sensitive data in etc/pulp/server/plugins.conf.d/*json
with ******* which is an invalid json value. Also it takes away a ","
and collected file is no longer a valid json. This change modifies the
regex used to get a valid json file.

Also Fixed Regex above it which was consuming any file having "conf"
even in middle of file path.

Closes #1878

Signed-off-by: Rohan Arora <roarora@redhat.com>
Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>