[java] Only run java conditionally and collect more
As this has been run unconditionaly to this point I choose to
do both commands = java, files=/usr/bin/java and packages.
There are many times Java is installed to /opt/ for say IBM's
java or Oracle's Java, or someone else's java.  To the same
 extent there are many different packages that include java
so do /usr/bin/java in case we miss some.

Add * to /etc/java*/ as some of the names can be different like
/etc/java-13-openjdk/

Remove alternatives from root symlink - I couldn't find a
reason for it to be in root, but happy to revisit.  (doesn't
work on Ubuntu)

Add java -version.  If it's in path we should at least be able
to get this.

Resolves: #1886

Signed-off-by: Bryan Quigley <bryan.quigley@canonical.com>