[memory] Collect slabtop info
Adds collection of `slabtop -o` to show the top consumers of SLAB.

Note that currently there is an issue with this command that will cause
it to hang if stdin does not receive a tty - which is exactly what
happens when commands are run via `timeout` without the `foreground`
option. As such, this command for now will use `foreground=True`
parameter for `add_cmd_output()`. Once/if this slaptop bug is fixed,
this plugin will be updated to no longer use that parameter.

Resolves: #1895

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>
Signed-off-by: Pavel Moravec <pmoravec@redhat.com>