[redhat] Add ability to upload to customer cases
Adds the ability to the Red Hat policy for users to upload archives
directly to existing technical support cases on the Red Hat Customer
Portal.

If a case number is provided, uploading to the case is the
preferred/default option the policy will take. If `--upload-user` is not
provided, users will be prompted for one as well as the password for the
Customer Portal account. Note that Red Hat's case management API
requires that the user credentials provided here have access to the case
number referenced on the Customer Portal.

If a case number is not provided, or if one is provided but Customer
Portal credentials are not provided, the policy will fallback to using
the public dropbox location for upload.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>