[utilities] Add --foreground to timeout commands optionally
A small number of recent issues have been identified that revolved
around certain commands hanging when run by sos unless `timeout=None`
was specified for those commands.

This patch adds `--foreground` to the `timeout` command that we use to
wrap any collected commands, so as to not disconnect the TTY which was
resulting in the hanging behavior. This addition is controlled by the
use of the new `foreground` parameter in any `add_cmd_output()` or
`collect_cmd_output()` call.

Resolves: #1909

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>