[Policies] Fix case id being overridden by blank input
Previously if `--case-id` was specific, and a user provided a blank
input for the case-id prompt (when `--batch` isn't used), this would
override the case id to be blank.

Now we properly default to `--caseid` if it is provided if a user
doesn't provide it again for the actual prompt.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>