[collectd] grab collectd.log whenever it is present
The plugin should be enabled also in a Super Privileged Container that
does not have the collectd package installed, but it has access and logs
to the host's collectd.log.

This logfile should be grabbed, alongside with config-data dir to ensure
the right configs were properly copied to /etc during deployment.

Resolves: #1917

Signed-off-by: Pavel Moravec <pmoravec@redhat.com>