[Plugin] Improve predicate failure message
Previously, the logged warning message when a predicate evaluation fails
would be a string that contained all elements passed to the predicate,
and include an empty list for checks that were not run by the predicate.
This meant that it was not immediately clear _which_ element(s) caused
the predicate evaluation to fail.

Now, during evaluation of the predicate checks, any failures are
recorded and the new failure message is generated based on these
failures. Thus, only the missing elements are presented to the user,
allowing for easier identification of what caused a particular
collection to be skipped.

Resolves: #1910

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>
Signed-off-by: Pavel Moravec <pmoravec@redhat.com>