[kubernetes|RedHat] Use `oc` command when running on OCP
If sos is run on a RHEL OpenShift node, use the `oc` command directly
instead of pointing `kubectl` at the origin config file. This allows sos
to not need to account for any other differences the `oc` binary might
implement when wrapping `kubectl`, which in turn could potentially lead
to a difference between what an end-user troubleshooting OCP may see and
what a support engineer looking through an sosreport may see.

Resolves: #1928

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>