[InitSystem|Plugin] Add method to get service names matching regex
Adds a new `get_service_names()` method to the `InitSystem` class which
is then exposed to plugins via `Plugin.get_service_names()` to allow
plugins to fetch a list of service names discovered on the system that
match a provided regex.

Included in this is a fix to avoid the header line from the discovery
command being entered in the dict of known services, so that we avoid
inadvertant regex matches.

Resolves: #1943

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>