[Juju] Juju Plugin Refactoring
Since the current plugin is having issues when we run sosreport
on a juju controller we have decided to refactor it and make
it simpler.

We had some discussions following three separate PRs:
https://github.com/sosreport/sos/pull/1670
https://github.com/sosreport/sos/pull/1671
https://github.com/sosreport/sos/pull/1672

We decided not to gather mongodb dumps as it will leak the
password from mongodb when running `ps`.

We decided to drop the output from juju commands as this
requires the command to be run as the user that holds the juju
credentials.

This plugin will gather logs from `/var/log/juju/*.log` and agent
configuration files from `/var/lib/juju/agents/*/agent.conf`, as
well as systemd logs related to the agents.
If sosreport is run with `--all-logs` it will gather all of the above
as well as all the files from `/var/lib/juju/` and `/var/log/juju`.

Closes: #1653

Co-authored-by: David Negreira david.negreira@canonical.com
Co-authored-by: Nick Niehoff nick.niehoff@canonical.com
Signed-off-by: David Negreira david.negreira@canonical.com