[yum] Add "yum -C repolist --verbose" command output
Following the upstream DNF commit below, the "yum -C repolist" command
output will no longer include package count information:

    c4f57476 Do not load metadata for repolist commands...

This information is helpful for those searching for client-side
mismatches to upstream repositories. Such as in the case of two
repositories, where one is missing content that another includes, this
information can be crucial for determinining the source of the problem.

This change includes a "yum -C repolist --verbose" command in the
resulting sosreport which includes information such as the following:

   Repo-id            : <repoid>
   Repo-name          : <full reponame>
   Repo-revision      : <version>
   Repo-updated       : <datestamp>
   Repo-pkgs          : <total repo package count>
   Repo-available-pkgs: <package count>
   Repo-size          : <size>
   Repo-baseurl       : <url>
   Repo-expire        : <datestamp>
   Repo-filename      : <path>

Signed-off-by: Kyle Walker <kwalker@redhat.com>