[Plugin] Allow negative matching for SoSPredicate
Extends SoSPredicate to allow for 'negative' presence checks, e.g. a
package is _not_ installed or a system is _not_ of a specified list of
architectures.

This is done by setting the relevant test in the 'required' dict handed
to SoSPredicate instances to be set to 'none' (the string, not the
NoneType).

For example, in order to run a command provided that a particular
service is not running, the following predicate would be used:

	SoSPredicate(self, services='foobar', required={'services':
'none'})

Related: #1975

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>