[Plugin] Do not write empty files to archive on command failure
If a command is not found on the system, sos traditionally will handle
that failed call by simply not writing any output to the archive.
However, if the command was given a suggested_filename or a root_symlink
parameter in the `add_cmd_output()` call, we would end up writing an
empty file to the archive.

Now, we will re-check the status of the command attempt after the chroot
check (and possible re-run) is done, and if the command "still" is not
found, exit immediately to avoid the empty writes.

Related: #1975

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>