[4.0] Create new binary and define new entrypoints for subcommands
This is the first patch to add a new `sos` binary to act as the new
commandline interface for users.

It adds the basic `SoS()` and `SoSComponent()` classes that will be
built out and used going forward for new subcommand development. Actual
functionality will be added in coming patches, as this is mainly serving
as a checkpoint to avoid over-large commits affecting wide swathes of
changes at once.

Related: #1986

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>