[options] Update option loading from config file
As we are moving to decouple SoSOptions() from report, we need to adjust
how some of the option loading is handled, such as loading values from a
conf file. The previous approach was to instantiate a new SoSOptions()
object, read into it from the config file, and then merge that with the
"main" SoSOptions() object. This worked well when we only had report to
worry about since the arguments were more or less static.

However, in allowing new subcommands to define their own options this
presented a problem due to the behavior of merging. This was compounded
by the fact that merging still functioned how we expected/needed for
actions like merging commandline options against the defaults.

As such, the new approach is to use the "main" SoSOptions() instance to
read the config file specifications and then directly modify the needed
values rather than relying on the merging behavior.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>